#37: returning old data model back to server, it will be handled later - now focusin only on the bear minimum
