--- model/src/main/scala/za/co/absa/atum/model/Flow.scala
@@ -1,30 +1,0 @@
-/*
- * Copyright 2021 ABSA Group Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package za.co.absa.atum.model
-
-import java.util.UUID
-
-case class Flow(
-  id: Option[UUID],
-  description: Option[String],
-  properties: Map[String, String] = Map.empty
-) extends BaseApiModel {
-
-  def apply(desc: Option[String] = None): Flow = Flow(None, desc)
-
-  override def withId(uuid: UUID): Flow = copy(id = Some(uuid))
-}
