+++ model/src/main/scala/za/co/absa/atum/model/Measurement.scala
@@ -17,8 +17,5 @@
 package za.co.absa.atum.model
 
 case class Measurement(
-  controlName: String,
-  controlType: String,
-  controlCol: String,
-  controlValue: Any
+    measurements: Map[Measure, String] = Map.empty
 )
