+++ server/src/main/scala/za/co/absa/atum/web/model/ControlMeasureMetadata.scala
@@ -1,0 +1,12 @@
+package za.co.absa.atum.web.model
+
+case class ControlMeasureMetadata(
+  sourceApplication: String,
+  country: String,
+  historyType: String,
+  dataFilename: String,
+  sourceType: String,
+  version: Int,
+  informationDate: String,
+  additionalInfo: Map[String, String] = Map.empty
+)
