+++ server/src/main/scala/za/co/absa/atum/web/model/Measurement.scala
@@ -1,0 +1,8 @@
+package za.co.absa.atum.web.model
+
+case class Measurement(
+  controlName: String,
+  controlType: String,
+  controlCol: String,
+  controlValue: Any
+)
